SCRUM-361: Send message of new matches (#19)
Send notifications to the bot for players whose reserve status changes to "Assigned" or transitions to "Assigned" status. When a match is generated, users are notified (without repetition) via a Telegram message. In this US, the new match message is created and sent to the bot with its Telegram ID.

Use cases:
- When a player's reserve status is set to "Assigned" (when creating a match), they should be notified of new matches.
- When a player's reserve status transitions to "Assigned" (from "Similar," for example), they should be notified of new matches.